Refactor: refine Manager.
