+++ server/src/server/definitionsManager.ts
@@ -1,11 +1,8 @@
-import { sync as glob } from "glob"
-import { DefinitionLink, Logger, URI, WorkspaceFolder } from "vscode-languageserver"
+import { DefinitionLink, URI } from "vscode-languageserver"
 import { TextDocument } from "vscode-languageserver-textdocument"
 
 import { parseDefinitions } from "@/services/definition"
-import { Settings } from "@/settings"
-import { disableLanguageServer } from "@/utilities/disableLanguageServer"
-import { makeDefinitionLinkMarkdown, readTextDocumentFromUri } from "@/utilities/text"
+import { makeDefinitionLinkMarkdown } from "@/utilities/text"
 
 export type Definition = string;
 export type DefinitionCandidate = {
@@ -14,13 +11,8 @@ export type DefinitionCandidate = {
 };
 
 export class DefinitionsManager {
-  private workspaceFolderUris: Set<URI> = new Set()
   private candidates: Map<Definition, DefinitionLink[]> = new Map()
   private fileDefinitions: Map<URI, Definition[]> = new Map()
-
-  hasWorkspaceFolder(workspaceFolder: WorkspaceFolder): boolean {
-    return this.workspaceFolderUris.has(workspaceFolder.uri)
-  }
 
   hasFileDefinitions(uri: URI): boolean {
     return this.fileDefinitions.has(uri)
@@ -47,41 +39,6 @@ export class DefinitionsManager {
     this.updateCandidates(document.uri, definitions)
 
     return definitions
-  }
-
-  async loadWorkspaceDefinitions(
-    workspaceFolder: WorkspaceFolder,
-    settings: Settings,
-    logger: Logger,
-  ): Promise<void> {
-    this.workspaceFolderUris.add(workspaceFolder.uri)
-
-    const files = [
-      ...new Set(
-        settings.definitionFiles.flatMap((filePattern) => glob(filePattern)),
-      ),
-    ]
-
-    for (const file of files) {
-      const documentUri = `${workspaceFolder.uri}/${file}`
-      const document = readTextDocumentFromUri(documentUri)
-
-      if (disableLanguageServer(document)) {
-        continue
-      }
-
-      try {
-        await this.updateFileDefinitions(
-          document, settings.defaultSchema,
-        )
-      }
-      catch (error: unknown) {
-        logger.error(
-          `The definitions of "${documentUri}" cannot load.`
-          + ` ${(error as Error).message}`,
-        )
-      }
-    }
   }
 
   private updateCandidates(
