+++ server/src/server/symbolsManager.ts
@@ -1,11 +1,7 @@
-import { sync as glob } from "glob"
-import { Logger, SymbolInformation, URI, WorkspaceFolder } from "vscode-languageserver"
+import { SymbolInformation, URI } from "vscode-languageserver"
 import { TextDocument } from "vscode-languageserver-textdocument"
 
 import { parseDocumentSymbols } from "@/services/symbol"
-import { Settings } from "@/settings"
-import { disableLanguageServer } from "@/utilities/disableLanguageServer"
-import { readTextDocumentFromUri } from "@/utilities/text"
 
 export type Definition = string;
 export type DefinitionCandidate = {
@@ -14,12 +10,7 @@ export type DefinitionCandidate = {
 };
 
 export class SymbolsManager {
-  private workspaceFolderUris: Set<URI> = new Set()
   private fileSymbols: Map<URI, SymbolInformation[]> = new Map()
-
-  hasWorkspaceFolder(workspaceFolder: WorkspaceFolder): boolean {
-    return this.workspaceFolderUris.has(workspaceFolder.uri)
-  }
 
   hasFileSymbols(uri: URI): boolean {
     return this.fileSymbols.has(uri)
@@ -41,40 +32,5 @@ export class SymbolsManager {
     this.fileSymbols.set(document.uri, symbols || [])
 
     return symbols
-  }
-
-  async loadWorkspaceSymbols(
-    workspaceFolder: WorkspaceFolder,
-    settings: Settings,
-    logger: Logger,
-  ): Promise<void> {
-    this.workspaceFolderUris.add(workspaceFolder.uri)
-
-    const files = [
-      ...new Set(
-        settings.definitionFiles.flatMap((filePattern) => glob(filePattern)),
-      ),
-    ]
-
-    for (const file of files) {
-      const documentUri = `${workspaceFolder.uri}/${file}`
-      const document = readTextDocumentFromUri(documentUri)
-
-      if (disableLanguageServer(document)) {
-        continue
-      }
-
-      try {
-        await this.updateFileSymbols(
-          document, settings.defaultSchema,
-        )
-      }
-      catch (error: unknown) {
-        logger.error(
-          `The symbols of "${documentUri}" cannot load.`
-          + ` ${(error as Error).message}`,
-        )
-      }
-    }
   }
 }
